--- js/utils/check-string-length.js
@@ -1,4 +1,0 @@
-function checkStringLength (string, length) {
-  return string.length <= length;
-}
-checkStringLength('asd', 2);
